+++ project_types.py
@@ -1,0 +1,3 @@
+from typing import Any
+
+NumpyArray = Any
